Update to LocalNotification.cancel()
Based on this answer from CommonsWare http://stackoverflow.com/a/16340703/2163901

When canceling a PendingIntent, there's no need to pass in extra flags to PendingIntent.getBroadcast. When you pass the flag in, the PendingIntent is not really cancelled. This can be confirmed by using the dumpsys utility as such: adb shell dumpsys alarm.