fix: wrap num in array object in getAll method
The `_exec method is wrapping non array arguments in an array object. However, the check that is utilized is `if (args)`. When this args variable contains the int 0, this will fail and no parameter will be passed like in line 364, in the method _getAll_.

Changed the other methods, _getScheduled_ and _getTriggered_, for the sake of consistency