fix: revert the concept of TRIGGERED notification (android)
With the merge of the upstream, the concept of the type of TRIGGERED notification was changed. According to the upstream, a TRIGGERED notification is a notification that is active in the Notification tray.

This behaviour was reverted to our previous definition of TRIGGERED -> a triggered notification is a notification that has a date in the past and is not recurrent.

References https://outsystemsrd.atlassian.net/browse/RNMT-3406