Creating intent with receiver when cancelling.
Fixes "too many alarms (500)" issue on Samsung devices after cancelling and
scheduling notifications, by creating the same Intent when cancelling an
existing notification, just like the one that was used to create the notification.
This clears both the PendingIntent and the scheduled Alarm.

Uses TriggerReceiver.class as the default receiver.

Use the "adb shell dumpsys alarm" command to confirm that the number of
alarms don't increase from your app, when cancelling and scheduling
notifications. Previously the Alarm remained in the alarm list when cancelling.