fix: don't use notification trampolin with Service for Android >= 12
Android 12 blocks the usage of notification trampolines. As so, we cannot delegate a Service to be opened when clicking on a notification, because that is a trampoline. Instead we create the PendingIntent using the MainActivity.

See: https://developer.android.com/about/versions/12/behavior-changes-12#notification-trampolines

References: https://outsystemsrd.atlassian.net/browse/RMET-822