Android: Use FLAG_IMMUTABLE on every Android version
Since Andorid 12 it is necessary to set the mutability of a PendingIntent. The flags for doing this are FLAG_IMMUTABLE or FLAG_MUTABLE, which exists already since Android 6, but the code was differentiating between Android versions if the flag should be set. The code should be consistent between all Android versions. It makes no sense to specify the mutability on Android 12 and newer, but not on older Anroid versions.